+++ src/Website/webpage.ts
@@ -1,13 +1,18 @@
-import { WebpageInterface, createWebpage } from './webpageInterface';
+import {WebpageInterface, createWebpage, validWebpageInterface} from './webpageInterface';
 
 export function webpageNote(info: WebpageInterface): string {
-  let webpage = createWebpage(info);
+  let webpage: validWebpageInterface = createWebpage(info);
   return `${webpage.authorList.note()}"${webpage.title}," ${webpage.nameOfSite}, ${publisherCitation(webpage.publisher)}${webpage.date}${isCommaRequired(webpage.publisher, webpage.date)}${webpage.url}.`;
 }
 
 export function webpageBibliography(info: WebpageInterface): string {
-  let webpage = createWebpage(info);
+  let webpage: validWebpageInterface = createWebpage(info);
   return `${webpage.authorList.bibliography()}"${webpage.title}." ${webpage.nameOfSite}. ${publisherCitation(webpage.publisher)}${webpage.date}${isPeriodRequired(webpage.publisher, webpage.date)}${webpage.url}.`;
+}
+
+export function webpageShortNote(info: WebpageInterface): string {
+  let webpage: validWebpageInterface = createWebpage(info);
+  return shortNotePrepend(webpage) + "\"" + webpage.title + ".\"";
 }
 
 function publisherCitation(publisher: string): string {
@@ -20,4 +25,8 @@ function isCommaRequired(pub: string, date: string): string {
 
 function isPeriodRequired(pub: string, date: string): string {
   return (pub === '' && date === '') ? '' : '. ';
+}
+
+function shortNotePrepend(citation: validWebpageInterface): string {
+  return (citation.authorList.length() > 0) ? citation.authorList.authors[0].last + ", " : citation.nameOfSite + ", ";
 }