+++ src/citation.ts
@@ -3,8 +3,8 @@ import { JournalInterface } from './Journal/journalInterface';
 import { WebpageInterface } from './Website/webpageInterface';
 import { LectureInterface } from './Lecture/lectureInterface';
 import { bookBibliography, bookNoteList } from './Book/book';
-import { webpageBibliography, webpageNote } from './Website/webpage';
-import { lectureBibliography, lectureNote } from './Lecture/lecture';
+import {webpageBibliography, webpageNote, webpageShortNote} from './Website/webpage';
+import {lectureBibliography, lectureNote, lectureShortNote} from './Lecture/lecture';
 import { journalBibliography, journalNoteList } from './Journal/journal';
 
 export function book(info: BookInterface, pages: string[]): {bibliography: string, notes: string[]} {
@@ -17,10 +17,14 @@ export function journal(info: JournalInterface, pages: {page: string}[]): {bibli
 
 export function webpage(info: WebpageInterface): {bibliography: string, notes: string[]} {
   let note: string = webpageNote(info);
-  return { bibliography: webpageBibliography(info), notes: [note] };
+  let shortNote: string = webpageShortNote(info);
+
+  return { bibliography: webpageBibliography(info), notes: [note, shortNote] };
 }
 
 export function lecture(info: LectureInterface): {bibliography: string, notes: string[]} {
   let note: string = lectureNote(info);
-  return { bibliography: lectureBibliography(info), notes: [note] };
+  let shortNote: string = lectureShortNote(info);
+
+  return { bibliography: lectureBibliography(info), notes: [note, shortNote] };
 }