+++ test/Website/webpage.test.ts
@@ -1,4 +1,4 @@
-import {webpageNote, webpageBibliography}  from "../../src/Website/webpage";
+import {webpageNote, webpageBibliography, webpageShortNote} from "../../src/Website/webpage";
 
 let webpage = {
     title: "Title of Webpage",
@@ -17,22 +17,40 @@ let noPub = {
     url: "wwww.testwebsite.com"
 };
 
+let noAuthor = {
+    title: "Title of Webpage",
+    publisher: "testPublisher",
+    nameOfSite: "testWebsite",
+    date: "2000",
+    url: "wwww.testwebsite.com"
+};
+
 describe("webpageNote() testing", () => {
-    test("should create note citation with all fields filled", () => {
+    it("should create note citation with all fields filled", () => {
         expect(webpageNote(webpage)).toBe("First Last, \"Title of Webpage,\" testWebsite, testPublisher, 2000, wwww.testwebsite.com.");
     });
 
-    test("should create note citation with missing publisher field", () => {
+    it("should create note citation with missing publisher field", () => {
         expect(webpageNote(noPub)).toBe("First Last, \"Title of Webpage,\" testWebsite, 2000, wwww.testwebsite.com.");
     });
 });
 
 describe("Webpage.bibliography() testing", () => {
-    test("should create bibliography citation with all fields filled", () => {
+    it("should create bibliography citation with all fields filled", () => {
         expect(webpageBibliography(webpage)).toBe("Last, First. \"Title of Webpage.\" testWebsite. testPublisher, 2000. wwww.testwebsite.com.");
     });
 
-    test("should create bibliography citation with missing publisher field", () => {
+    it("should create bibliography citation with missing publisher field", () => {
         expect(webpageBibliography(noPub)).toBe("Last, First. \"Title of Webpage.\" testWebsite. 2000. wwww.testwebsite.com.");
+    });
+});
+
+describe("Webpage.shortNote() testing", () => {
+    it("should create short note citation with all fields filled", () => {
+        expect(webpageShortNote(webpage)).toBe("Last, \"Title of Webpage.\"");
+    });
+
+    it("should create short note with missing author", () => {
+        expect(webpageShortNote(noAuthor)).toBe("testWebsite, \"Title of Webpage.\"");
     });
 });